added blog page that includes blog previews, changed homepage
