+++ my-app/src/app/api/[slug]/route.ts
