--- my-app/src/components/footer.module.css
@@ -1,2 +1,0 @@
-
-
