+++ my-app/src/components/footer.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import style from "./navbar.module.css";
 
 export default function Footer() {
     return(
